pyproject.toml
[project]
name = "capstone-model-and-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.8",
    "jupyter>=1.1.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "optuna>=4.3.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "shap>=0.47.2",
    "smartsheet-python-sdk>=3.0.5",
    "torch>=2.7.0",
]
